tests: Add more graphics tests
These tests several vkCmdDraw()s within a single dynamic
rendering block, only one of them is an infinite loop. Ideally
we should be able to have the hang region only include one draw call.
